Move constructors into their associated classes
Lexicon.cs, ListValue.cs, PIDLoop.cs, QueueValue.cs
StackValue.cs, UniqueSetValue.cs, PathValue.cs, Volume.cs
* Added nested classes from Misc.cs, Persistence.cs, and Suffixed.cs to
host kerboscript constructor functions within the associated structure

STYLEGUIDE.md
* Update style guide to specifically call out acceptability of nested
classes and preference in the case of kerboscript constructor classes

Also som minor white space modifications
