+++ STYLEGUIDE.md
@@ -2,8 +2,8 @@
 
 ## Definitions
 
-* [CamelCase](http://en.wikipedia.org/wiki/CamelCase) is a casing convention where the first letter is lower-case, words are not separated by any character but have their first letter capitalized. Example: <code>thisIsCamelCased</code>. 
-* [PascalCase](http://c2.com/cgi/wiki?PascalCase) is a casing convention where the first letter of each word is capitalized, and no separating character is included between words. Example: <code>ThisIsPascalCased</code>. 
+* [CamelCase](http://en.wikipedia.org/wiki/CamelCase) is a casing convention where the first letter is lower-case, words are not separated by any character but have their first letter capitalized. Example: <code>thisIsCamelCased</code>.
+* [PascalCase](http://c2.com/cgi/wiki?PascalCase) is a casing convention where the first letter of each word is capitalized, and no separating character is included between words. Example: <code>ThisIsPascalCased</code>.
 
 ## C# coding conventions
 
@@ -11,11 +11,31 @@ We should use the [Allman bracing style](http://en.wikipedia.org/wiki/Indent_sty
 
 We are using the C# coding conventions described in this document as a guide, not everything in this doc is gospel and is open to debate: [C# Coding Guidelines](http://blogs.msdn.com/brada/articles/361363.aspx) with the following exceptions:
 
-* Each file should not start with a copyright notice. The ones at the root of the source tree will suffice. 
-* Regions (#region) are not used. 
-* using statements are on top of a file (outside of namespace {...}) 
-* Use var only if you have an anonymous type or you can clearly tell what the type is from the right hand side of the expression 
+* Each file should not start with a copyright notice. The ones at the root of the source tree will suffice.
+* Regions (#region) are not used.
+* using statements are on top of a file (outside of namespace {...})
+* Use var only if you have an anonymous type or you can clearly tell what the type is from the right hand side of the expression
 * Member variables should always be private, public access should be provided by an encapsulated property.
+* While each source file should be limited to a single public class, nested
+  classes applicable to only their parent class are permited.
+* In the case of defining a kerboscript function class to serve as a constructor
+  for a specific structure, this function class should be nested at the top of
+  the structure it returns.  For example:
+  ```
+      public class ExampleStructure : Structure
+	  {
+		  [Function("example")] // insert the function at the top of the class
+		  public class FunctionExampleStructureConstructor
+		  {
+			  public override void Execute()
+			  {
+				  // insert logic here
+			  }
+		  }
+
+		  // insert member variables/constructors/methods per the style guide
+	  }
+  ```
 
 #### Naming
 Follow all .NET Framework Design Guidelines for both internal and external members. Highlights of these include:
@@ -69,7 +89,7 @@ Here is some sample code that follows these conventions.
 				}
 				return String.Empty;
 			}
-			
+
 			private string AnotherMethod(){
 				return privateMember.Count;
 			}
